a2a/weather_service/pyproject.toml
[project]
name = "weather-service"
version = "0.0.1"
description = "Simple ollama-based Langgraph agent with MCP tool calling."
authors = [
    { name = "Paolo Dettori" }
]
readme = "README.md"
license = { text = "Apache" }
requires-python = ">=3.11"
dependencies = [
    "a2a-sdk>=0.2.16",
    "langgraph>=0.2.55",
    "langchain-community>=0.3.9",
    "langchain-ollama>=0.2.1",
    "langchain-openai>=0.3.7",
    "openinference-instrumentation-langchain>=0.1.27",
    "pydantic-settings>=2.8.1",
    "langchain-mcp-adapters>=0.1.0",
    "python-keycloak>=5.5.1",
    "opentelemetry-exporter-otlp",
    # GenAI semantic convention instrumentation for token metrics
    "opentelemetry-instrumentation-openai>=0.34b0",
]

[project.scripts]
server = "weather_service.agent:run"


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
